Alterations in names and directories of some arquives.
